+++ lib/dns_forward_lwt_unix.ml
@@ -456,13 +456,31 @@ module Time = struct
   let sleep = Lwt_unix.sleep
 end
 
+module Clock = struct
+  type tm =
+    { tm_sec: int;
+      tm_min: int;
+      tm_hour: int;
+      tm_mday: int;
+      tm_mon: int;
+      tm_year: int;
+      tm_wday: int;
+      tm_yday: int;
+      tm_isdst: bool;
+    }
+
+  let time = Unix.gettimeofday
+
+  let gmtime _ = failwith "gmtime unimplemented"
+end
+
 module R = struct
   open Dns_forward
   module Udp_client = Rpc.Client.Make(Udp)(Framing.Udp(Udp))(Time)
-  module Udp = Resolver.Make(Udp_client)(Time)
+  module Udp = Resolver.Make(Udp_client)(Time)(Clock)
 
   module Tcp_client = Rpc.Client.Make(Tcp)(Framing.Tcp(Tcp))(Time)
-  module Tcp = Resolver.Make(Tcp_client)(Time)
+  module Tcp = Resolver.Make(Tcp_client)(Time)(Clock)
 end
 
 module Server = struct