+++ lib/dns_forward_resolver.ml
@@ -61,7 +61,6 @@ let choose_servers config request =
       (fun order ->
         List.filter (fun server -> server.Server.order = order) all
       ) (IntSet.elements orders)
-    |> List.concat
   | _ -> []
   end
 
@@ -72,32 +71,48 @@ let or_fail_msg m = m >>= function
 
 module type S = Dns_forward_s.RESOLVER
 
-module Make(Client: Dns_forward_s.RPC_CLIENT)(Time: V1_LWT.TIME) = struct
+module Make(Client: Dns_forward_s.RPC_CLIENT)(Time: V1_LWT.TIME)(Clock: V1.CLOCK) = struct
 
   module Cache = Dns_forward_cache.Make(Time)
 
   type address = Dns_forward_config.Address.t
   type message_cb = ?src:address -> ?dst:address -> buf:Cstruct.t -> unit -> unit Lwt.t
 
+  type connection = {
+    server: Dns_forward_config.Server.t;
+    client: Client.t;
+    mutable reply_expected_since: float option;
+    (* if None: we don't expect a reply
+       if Some t: we haven't heard from the server since time t *)
+    mutable replies_missing: int;
+    (* the number of requests we've sent which have not been replied to *)
+    mutable online: bool;
+    (* true if we assume the server is online *)
+  }
+
   type t = {
-    connections: (Dns_forward_config.Server.t * Client.t) list;
+    connections: connection list;
     local_names_cb: (Dns.Packet.question -> Dns.Packet.rr list option Lwt.t);
     cache: Cache.t;
+    config: Dns_forward_config.t;
   }
 
   let create ?(local_names_cb=fun _ -> Lwt.return_none) ?message_cb config =
     Lwt_list.map_s (fun server ->
       or_fail_msg @@ Client.connect ?message_cb server.Dns_forward_config.Server.address
       >>= fun client ->
-      Lwt.return (server, client)
+      let reply_expected_since = None in
+      let replies_missing = 0 in
+      let online = true in
+      Lwt.return { server; client; reply_expected_since; replies_missing; online }
     ) (Dns_forward_config.Server.Set.elements config.Dns_forward_config.servers)
     >>= fun connections ->
     let cache = Cache.make () in
-    Lwt.return { connections; local_names_cb; cache }
+    Lwt.return { connections; local_names_cb; cache; config }
 
   let destroy t =
     Cache.destroy t.cache;
-    Lwt_list.iter_s (fun (_, c) -> Client.disconnect c) t.connections
+    Lwt_list.iter_s (fun c -> Client.disconnect c.client) t.connections
 
   let answer buffer t =
     let len = Cstruct.len buffer in
@@ -135,20 +150,52 @@ module Make(Client: Dns_forward_s.RPC_CLIENT)(Time: V1_LWT.TIME) = struct
             match Cache.answer t.cache address question with
             | Some answers -> Lwt.return (Ok (`Success (marshal_reply answers)))
             | None ->
-              let _, client = List.find (fun (s, _) -> s = server) t.connections in
+              let c = List.find (fun c -> c.server = server) t.connections in
               begin
+                let now = Clock.time () in
                 (* If no timeout is configured, we will stop listening after
                    5s to avoid leaking threads if a server is offline *)
                 let timeout_ms = match server.Server.timeout_ms with None -> 5000 | Some x -> x in
-                Lwt.pick [
-                  ( Time.sleep (float_of_int timeout_ms /. 1000.0)
-                    >>= fun () ->
-                    Lwt.return (Error (`Msg "timeout")) );
-                  Client.rpc client buffer
-                ]
+                (* If no assume_offline_after_drops is configured then set this
+                   to 5s. *)
+                let assume_offline_after_drops = match t.config.assume_offline_after_drops with
+                  | Some c -> c
+                  | None -> 5 in
+                (* Within the overall timeout_ms (configured by the user) we will send
+                   the request at 1s intervals to guard against packet drops. *)
+                let delays_ms =
+                  let rec make from = if from > timeout_ms then [] else from :: (make (from + 1000)) in
+                  make 0 in
+                let requests = List.map (fun delay_ms ->
+                  Time.sleep (float_of_int delay_ms /. 1000.0)
+                  >>= fun () ->
+                  Client.rpc c.client buffer
+                ) delays_ms in
+                let timeout =
+                  Time.sleep (float_of_int timeout_ms /. 1000.0)
+                  >>= fun () ->
+                  Lwt.return (Error (`Msg "timeout")) in
+                Lwt.pick (timeout :: requests)
                 >>= function
-                | Error x -> Lwt.return (Error x)
+                | Error x ->
+                  if c.reply_expected_since = None then c.reply_expected_since <- Some now;
+                  c.replies_missing <- c.replies_missing + (List.length delays_ms);
+                  if assume_offline_after_drops < c.replies_missing then begin
+                    Log.err (fun f -> f "Upstream DNS server %s has dropped %d packets in a row: assuming it's offline"
+                      (Dns_forward_config.Address.to_string address) c.replies_missing
+                    );
+                    c.online <- false
+                  end;
+                  Lwt.return (Error x)
                 | Ok reply ->
+                  c.reply_expected_since <- None;
+                  c.replies_missing <- 0;
+                  if not c.online then begin
+                    Log.info (fun f -> f "Upstream DNS server %s is back online"
+                      (Dns_forward_config.Address.to_string address)
+                    );
+                    c.online <- true;
+                  end;
                   (* Determine whether it's a success or a failure; if a success
                      then insert the value into the cache. *)
                   let len = Cstruct.len reply in
@@ -162,43 +209,76 @@ module Make(Client: Dns_forward_s.RPC_CLIENT)(Time: V1_LWT.TIME) = struct
                   end
               end in
 
-          (* Filter the list of servers using any "zone" setting -- this will
-             prevent queries for private names being leaked to public servers
-             (if configured).
-             Group the servers into lists of equal priorities. *)
-          choose_servers (List.map fst t.connections) request
-          |>
-          (* Send all requests in parallel to minimise the chance of hitting a
-             timeout. Positive replies will be cached, but servers which don't
-             recognise the name will be queried each time. *)
-          List.map one_rpc
-          |>
+          (* Ask many servers but first
+             - Filter the list of servers using any "zone" setting -- this will
+               prevent queries for private names being leaked to public servers
+               (if configured).
+             - Group the servers into lists of equal priorities.
+             - Send all the requests concurrently. *)
+          let many_rpcs connections =
+            let equal_priority_groups = choose_servers (List.map (fun c -> c.server) connections) request  in
+            (* Send all requests in parallel to minimise the chance of hitting a
+               timeout. Positive replies will be cached, but servers which don't
+               recognise the name will be queried each time. *)
+            List.map (List.map one_rpc) equal_priority_groups in
+
+          let online, offline = List.partition (fun c -> c.online) t.connections in
+          if online = [] then begin
+            let open Dns_forward_config in
+            Log.warn (fun f -> f "There are no online DNS servers configured.");
+            Log.warn (fun f -> f "DNS servers %s are all marked offline"
+              (String.concat ", " (List.map (fun c -> Address.to_string @@ c.server.Server.address) offline))
+            )
+          end;
+          (* For all the offline servers, send the requests as a "ping" to see
+             if they are alive or not. Any response will flip them back to online
+             but we won't consider their responses until the next RPC *)
+          let _ = many_rpcs offline in
+
+          (* For all the online servers, send the requests and return the waiting
+             threads. *)
+          let online_results = many_rpcs online in
+
+          (* Wait for the best result from a set of equal priority requests *)
+          let rec wait best_so_far remaining =
+            if remaining = []
+            then Lwt.return best_so_far
+            else
+              Lwt.nchoose_split remaining
+              >>= fun (terminated, remaining) ->
+              match List.fold_left
+                (fun best_so_far next -> match best_so_far with
+                  | Ok (`Success result) ->
+                    (* No need to wait for any of the rest: one success is good enough *)
+                    Ok (`Success result)
+                  | best_so_far ->
+                    begin match best_so_far, next with
+                    | _, Ok (`Success result) ->
+                      Ok (`Success result)
+                    | Ok (`Failure (a_packet, a_reply)), Ok (`Failure (b_packet, b_reply)) ->
+                      begin match a_packet, b_packet with
+                      (* Prefer NXDomain to errors like Refused *)
+                      | Some { detail = { rcode = NXDomain; _ }; _ }, _ -> Ok (`Failure (a_packet, a_reply))
+                      | _, Some { detail = { rcode = NXDomain; _ }; _ } -> Ok (`Failure (b_packet, b_reply))
+                      | _, _ ->
+                        (* other than that, the earlier error is better *)
+                        Ok (`Failure (a_packet, a_reply))
+                      end
+                    | Error _, Ok (`Failure (b_packet, b_reply)) ->
+                      (* prefer a high-level error over a low-level (e.g. socket) error *)
+                      Ok (`Failure (b_packet, b_reply))
+                    | best_so_far, _ ->
+                      best_so_far
+                    end
+                ) best_so_far terminated with
+              | Ok (`Success result) -> Lwt.return (Ok (`Success result))
+              | best_so_far -> wait best_so_far remaining in
+          (* Wait for each equal priority group at a time *)
           Lwt_list.fold_left_s
-            (fun best_so_far next_t -> match best_so_far with
-              | Ok (`Success result) ->
-                (* No need to wait for any of the rest: one success is good enough *)
-                Lwt.return (Ok (`Success result))
-              | best_so_far ->
-                next_t >>= fun next ->
-                begin match best_so_far, next with
-                | _, Ok (`Success result) ->
-                  Lwt.return (Ok (`Success result))
-                | Ok (`Failure (a_packet, a_reply)), Ok (`Failure (b_packet, b_reply)) ->
-                  begin match a_packet, b_packet with
-                  (* Prefer NXDomain to errors like Refused *)
-                  | Some { detail = { rcode = NXDomain; _ }; _ }, _ -> Lwt.return (Ok (`Failure (a_packet, a_reply)))
-                  | _, Some { detail = { rcode = NXDomain; _ }; _ } -> Lwt.return (Ok (`Failure (b_packet, b_reply)))
-                  | _, _ ->
-                    (* other than that, the earlier error is better *)
-                    Lwt.return (Ok (`Failure (a_packet, a_reply)))
-                  end
-                | Error _, Ok (`Failure (b_packet, b_reply)) ->
-                  (* prefer a high-level error over a low-level (e.g. socket) error *)
-                  Lwt.return (Ok (`Failure (b_packet, b_reply)))
-                | best_so_far, _ ->
-                  Lwt.return best_so_far
-                end
-            ) (Error (`Msg "no servers configured"))
+            (fun best_so_far next -> match best_so_far with
+              | Ok (`Success result) -> Lwt.return (Ok (`Success result))
+              | best_so_far -> wait best_so_far next
+            ) (Error (`Msg "no servers configured")) online_results
           >>= function
           | Ok (`Success reply) -> Lwt_result.return reply
           | Ok (`Failure (_, reply)) -> Lwt_result.return reply